Refactor repo handling code into a new file, and rethink Links.
